% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterfly-data.R
\docType{data}
\name{butterfly}
\alias{butterfly}
\title{Butterfly data}
\format{
An object of class \code{"matrix"}.
}
\source{
\href{}{}
}
\usage{
data("butterfly")
}
\description{
Butterfly data
}
\examples{
data("butterfly")
}
\references{
Wang et al., 2018
}
\keyword{datasets}
