% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterflyNE-data.R
\docType{data}
\name{butterflyNE}
\alias{butterflyNE}
\title{Butterfly data after Network Enhancment with orignal Matlab implementation}
\format{
An object of class \code{"matrix"}.
}
\source{
\href{}{}
}
\usage{
data("butterflyNE")
}
\description{
Butterfly data after Network Enhancment with orignal Matlab implementation
}
\examples{
data("butterflyNE")
}
\references{
Wang et al., 2018
}
\keyword{datasets}
