% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominateset.R
\name{dominateset}
\alias{dominateset}
\title{dominateset}
\usage{
dominateset(aff_matrix, NR_OF_KNN)
}
\arguments{
\item{NR_OF_KNN}{- number of neighbors}

\item{affy_matrix}{- input weighted N x N adjacency matrix}
}
\value{
None
}
\description{
R implementation of Network Enhacement
}
\references{
\href{https://bit.ly/2RS5pLX}{Wang, Pourshafeie, Zitnik, et al., 2018}
}
\author{
danro9685, \github{https://github.com/danro9685/CIMLR}

Bo Wang, \github{https://github.com/wangboyunze}

microbma, \github{https://github.com/microbma}

Tyler Bradshaw, \github{https://github.com/twesleyb}
}
\keyword{enhancement}
\keyword{network}
