% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlab_Length.R
\name{Length}
\alias{Length}
\title{Length}
\usage{
Length(dm)
}
\arguments{
\item{dm}{- input matrix}
}
\value{
largest dimension of a matrix
}
\description{
R implementation of matlab's length function
}
\references{
\href{https://bit.ly/2RS5pLX}{Wang, Pourshafeie, Zitnik, et al., 2018}
}
\author{
microbma, \href{https://github.com/microbma}{github}

Tyler Bradshaw, \href{https://github.com/twesleyb}{github}
}
\keyword{enhancement}
\keyword{network}
