% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_Enhancement.R
\name{neten}
\alias{neten}
\title{neten}
\usage{
neten(W_in, weight = "weight", alpha = 0.9, diffusion = 2,
  k = NULL)
}
\arguments{
\item{W_in}{- input weighted N x N adjacency matrix}

\item{alpha}{-}

\item{diffusion}{- the extent of diffusion}

\item{k}{- number of neighbors}
}
\value{
re-weighted N x N adjacency matrix.
}
\description{
R implementation of Network Enhancement
}
\examples{
data(butterfly)
neten(butterfly)
}
\references{
\href{https://bit.ly/2RS5pLX}{Wang, Pourshafeie, Zitnik, et al., 2018}
}
\author{
Bo Wang, \href{https://github.com/wangboyunze}{github}

microbma, \href{https://github.com/microbma}{github}

Tyler Bradshaw, \href{https://github.com/twesleyb}{github}
}
\keyword{enhancement}
\keyword{network}
