% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlab_dominateset.R
\name{dominateset}
\alias{dominateset}
\title{dominateset}
\usage{
dominateset(aff_matrix, NR_OF_KNN)
}
\arguments{
\item{NR_OF_KNN}{- number of neighbors}

\item{affy_matrix}{- input weighted N x N adjacency matrix}
}
\value{
None
}
\description{
R implementation of Network Enhacement
}
\references{
\href{https://bit.ly/2RS5pLX}{Wang, Pourshafeie, Zitnik, et al., 2018}
}
\author{
danro9685, \href{https://github.com/danro9685/CIMLR}{github}

Bo Wang, \href{https://github.com/wangboyunze}{github}

microbma, \href{https://github.com/microbma}{github}

Tyler Bradshaw, \href{https://github.com/twesleyb}{github}
}
\keyword{enhancement}
\keyword{network}
