% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NE_dn.R
\name{NE_dn}
\alias{NE_dn}
\title{NE_dn}
\usage{
NE_dn(w, type = "avg")
}
\arguments{
\item{w}{- input weighted N x N adjacency matrix}

\item{type}{- normalization method to be used, defaults to 'avg' for average (aka 'ave')}
}
\value{
normalized adjacency matrix
}
\description{
R implementation of Network Enhancement
}
\references{
\href{https://bit.ly/2RS5pLX}{Wang, Pourshafeie, Zitnik, et al., 2018}
}
\author{
Bo Wang, \href{https://github.com/wangboyunze}{github}

microbma, \href{https://github.com/microbma}{github}

Tyler Bradshaw, \href{https://github.com/twesleyb}{github}
}
