% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhanceNetwork.R
\name{neten}
\alias{neten}
\title{neten}
\usage{
neten(W_in, alpha = 0.9, diffusion = 2, k = NULL)
}
\arguments{
\item{W_in}{- input weighted N x N adjacency matrix}

\item{alpha}{- regularization parameter}

\item{diffusion}{- diffusion parameter, typical values are 0.5, 1.0, 1.2, 2.0}

\item{k}{- number of neighbors, if NULL then defaults to ceil(20,ncol(data))}
}
\value{
re-weighted N x N adjacency matrix
}
\description{
R implementation of Network Enhancement
}
\examples{
library(neten)
data(butterfly)
neten(butterfly)
}
\references{
\href{https://bit.ly/2RS5pLX}{Wang et al., 2018}
}
\author{
Bo Wang, \href{https://github.com/wangboyunze}{github}

microbma, \href{https://github.com/microbma}{github}

Tyler Bradshaw, \href{https://github.com/twesleyb}{github}
}
\keyword{enhancement}
\keyword{network}
