% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Network_Enhancement.R
\name{Network_Enhancement}
\alias{Network_Enhancement}
\title{Network_Enhancement}
\usage{
Network_Enhancement(W_in, alpha = 0.9, diffusion = 2, k = min(20,
  ceiling(Length(W_in)/10)))
}
\arguments{
\item{W_in}{- input weighted N x N adjacency matrix}

\item{alpha}{-}

\item{diffusion}{- the extent of diffusion}

\item{k}{- number of neighbors}
}
\value{
re-weighted N x N adjacency matrix.
}
\description{
R implementation of Network Enhancement
}
\examples{
data(butterfly)
Network_Enhancement(butterfly)
}
\references{
\href{https://bit.ly/2RS5pLX}{Wang, Pourshafeie, Zitnik, et al., 2018}
}
\author{
Bo Wang, \github{https://github.com/wangboyunze}

microbma, \github{https://github.com/microbma}

Tyler Bradshaw, \github{https://github.com/twesleyb}
}
\keyword{enhancement}
\keyword{network}
