% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransitionFields.R
\name{TransitionFields}
\alias{TransitionFields}
\title{TransitionFields}
\usage{
TransitionFields(W)
}
\arguments{
\item{w}{- input weighted N x N adjacency matrix}

\item{diffusion}{- the extent of diffusion}
}
\value{
re-weighted N x N adjacency matrix.
}
\description{
R implementation of Network Enhancement
}
\references{
\href{https://bit.ly/2RS5pLX}{Wang, Pourshafeie, Zitnik, et al., 2018}
}
\author{
Bo Wang, \href{https://github.com/wangboyunze}{github}

microbma, \href{https://github.com/microbma}{github}

Tyler Bradshaw, \href{https://github.com/twesleyb}{github}
}
\keyword{enhancement}
\keyword{network}
